requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-auto-logout==0.5.0
django-filter==21.1
django-mathfilters==1.0.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==4.6.11
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
zipp==3.7.0
